Support for masked lt fringe + lt main kernels. (#217)
This PR introduces support for generating masked lt fringe and lt main
kernels. Currently the remainder n value (n<NR or NR+lt16) is computed as
a combination of a fringe/main kernel and a single lt kernel. Now new lt
kernels are generated which can enable the remainder computation as one
lt fringe call, with the fringe kernel dimension higher than it used to
be previously. Currently this is supported when generating AVX512 ZMM
instruction favoring kernels.
-A mechanism to configure the terminal fringe NR is added. For example,
now for AVX512 machine, we can have lt48 as the final fringe kernel
instead of lt16.
-Updates to kernelInfo and DE to enable this workflow.

AMD-Internal: [CPUPL-7525]

Signed-off-by: Mithun Mohan <MithunMohan.KadavilMadanaMohanan@amd.com>